fix: improve YAML preview formatting in web UI
Move addStepSpacing to shared src/helpers.ts so both CLI (generator.ts)
and web (lib.ts) apply consistent blank-line spacing between steps.
Remove dead no-op formattedYaml wrapper in yaml-preview.tsx.
Exclude .claude/worktrees from Jest to prevent stale worktree failures.
